finish price pairs dropdown
